pyproject.toml
[project]
name = "cruxpy"
description = "Help building a webpage for Crux ports repository"
license = "GPL-3.0-only"
version = "0.1.1"
dependencies = [
  "GitPython"
]
requires-python= ">=3.10"
authors = [
  { name = "Wawrzyniec Niewodniczański", email = "main@wawrzek.name" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Topic :: System :: Installation/Setup",
  "Topic :: System :: Software Distribution",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
]

[project.urls]
repository = "https://github.com/wawrzek/cruxpy"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
cruxpy= ["files/*"]
